trying to use sphinx to generate some documentation
