+++ ctoybox/ctoybox/ffi.py
@@ -2,87 +2,7 @@ from .ctoybox import lib, ffi
 import numpy as np
 from PIL import Image
 import json
-
-class Input():
-    """An input object represents a game controller having left, right, up, down, and two buttons.
-
-    ALE mapping:
-            ALE_ACTION_MEANING = {
-            0 : "NOOP",
-            1 : "FIRE",
-            2 : "UP",
-            3 : "RIGHT",
-            4 : "LEFT",
-            5 : "DOWN",
-            6 : "UPRIGHT",
-            7 : "UPLEFT",
-            8 : "DOWNRIGHT",
-            9 : "DOWNLEFT",
-            10 : "UPFIRE",
-            11 : "RIGHTFIRE",
-            12 : "LEFTFIRE",
-            13 : "DOWNFIRE",
-            14 : "UPRIGHTFIRE",
-            15 : "UPLEFTFIRE",
-            16 : "DOWNRIGHTFIRE",
-            17 : "DOWNLEFTFIRE",
-        }
-    """
-
-    _LEFT = "left"
-    _RIGHT = "right"
-    _UP = "up"
-    _DOWN = "down"
-    _BUTTON1 = "button1"
-    _BUTTON2 = "button2"
-    _NOOP = "noop"
-
-    def __init__(self):
-        self.reset()
-
-    def reset(self):
-        self.left = False
-        self.right = False
-        self.up = False
-        self.down = False
-        self.button1 = False
-        self.button2 = False
-
-    def __str__(self):
-        return self.__dict__.__str__()
-
-    def __repr__(self):
-        return self.__dict__.__str__()
-
-    def set_input(self, input_dir, button=_NOOP):
-        input_dir = input_dir.lower()
-        button = button.lower()
-
-        # reset all directions
-        if   input_dir == Input._NOOP:
-            pass
-        elif input_dir == Input._LEFT:
-            self.left = True
-        elif input_dir == Input._RIGHT:
-            self.right = True
-        elif input_dir == Input._UP:
-            self.up = True
-        elif input_dir == Input._DOWN:
-            self.down = True
-        else:
-            print('input_dir:', input_dir)
-            assert False
-
-        # reset buttons
-        if button == Input._NOOP:
-            pass
-        elif button == Input._BUTTON1:
-            self.button1 = True
-        elif button == Input._BUTTON2:
-            self.button2 = True
-        else:
-            assert False
-
+from .input import Input
 
 def rust_str(result):
     """
