+++ ctoybox/source/index.rst
@@ -1,0 +1,33 @@
+.. ctoybox documentation master file, created by
+   sphinx-quickstart on Fri Jan  3 16:52:02 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to ctoybox's documentation!
+===================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   input
+   simulator
+   state
+   toybox
+
+
+.. automodule:: ctoybox
+  :members:
+  :undoc-members:
+
+
+.. automodule:: ctoybox.ffi
+  :members:
+  :undoc-members:
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
