+++ ctoybox/source/input.rst
@@ -1,0 +1,6 @@
+Input
+=====
+
+.. automodule:: ctoybox.Input
+  :members:
+  :undoc-members:
