+++ ctoybox/source/simulator.rst
@@ -1,0 +1,7 @@
+Simulator
+=========
+
+.. autoclass:: ctoybox.Simulator
+  :members:
+  :undoc-members:
+
