+++ ctoybox/source/state.rst
@@ -1,0 +1,6 @@
+State
+=====
+
+.. autoclass:: ctoybox.State
+  :members:
+  :undoc-members:
