+++ ctoybox/source/toybox.rst
@@ -1,0 +1,6 @@
+Toybox
+======
+
+.. autoclass:: ctoybox.Toybox
+  :members:
+  :undoc-members:
